backend/requirements.txt
langchain
lagchain-community
langchain-google-genai
python-dotenv
fastapi
pypdf